+++ src/main/kotlin/lotto/domain/Rank.kt
@@ -1,6 +1,6 @@
 package lotto.domain
 
-enum class Rank(val hitCount: Int, val prize: Int) {
+enum class Rank(val hitCount: Int, val prize: Long) {
     ThreeHit(3, 5_000),
     FourHit(4, 50_000),
     FiveHit(5, 1_500_000),
